requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyAudio==0.2.14
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3
